Update files and folders
