+++ src/App.jsx
@@ -1,70 +1,14 @@
-import {
-  createBrowserRouter,
-  createRoutesFromElements,
-  Route,
-  RouterProvider,
-} from "react-router";
+import theme from "./assets/styles/theme";
+import AppRouter from "./routes/AppRouter";
 import CssBaseline from "@mui/material/CssBaseline";
-import { createTheme, ThemeProvider } from "@mui/material/styles";
-import { yellow, grey } from "@mui/material/colors";
-import MainLayout from "./layout/MainLayout";
-import NotFoundPage from "./pages/NotFoundPage";
-import HomePage from "./pages/HomePage";
-import AboutPage from "./pages/AboutPage";
-import ContactPage from "./pages/ContactPage";
-
-const router = createBrowserRouter(
-  createRoutesFromElements(
-    <Route path="/" element={<MainLayout />}>
-      <Route path="*" element={<NotFoundPage />} />
-      <Route index element={<HomePage />} />
-      <Route path="/about" element={<AboutPage />} />
-      <Route path="/contact" element={<ContactPage />} />
-    </Route>
-  )
-);
-
-const theme = createTheme({
-  palette: {
-    primary: {
-      main: grey[50],
-    },
-    secondary: {
-      main: yellow[800],
-    },
-  },
-  typography: {
-    fontFamily: "Roboto, sans-serif",
-    h1: {
-      fontFamily: "Ranchers, cursive",
-    },
-    h2: {
-      fontFamily: "Ranchers, cursive",
-    },
-    h3: {
-      fontFamily: "Ranchers, cursive",
-    },
-    h4: {
-      fontFamily: "Ranchers, cursive",
-    },
-    h5: {
-      fontFamily: "Ranchers, cursive",
-    },
-    h6: {
-      fontFamily: "Ranchers, cursive",
-    },
-    button: {
-      fontFamily: "Ranchers, cursive",
-    },
-  },
-});
+import { ThemeProvider } from "@mui/material/styles";
 
 function App() {
   return (
     <>
       <CssBaseline />
       <ThemeProvider theme={theme}>
-        <RouterProvider router={router} />
+        <AppRouter />
       </ThemeProvider>
     </>
   );
